Split the file browser/uploader into its own component to allow for reuse
